bump 310

+++ core/cairo_wrapper.h
@@ -203,8 +203,8 @@ namespace cairo_wrapper
 
     struct BufferView {
         const char* data;
-        ssize_t size;
-        ssize_t offset;
+        unsigned int size;
+        unsigned int offset;
     };
 
     cairo_status_t write_to_vector(void* closure, const unsigned char* data, unsigned int length);
